objId is size_t and compulsory
git-svn-id: file:///home/roberto/Git/svn@7156 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0e383d0bc72c9f89b40a1d5112f0ca6ecddbc044

+++ trunk/xmipp/applications/programs/statistics/statistics_main.cpp
@@ -100,7 +100,7 @@ protected:
         }
     }
 
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
     {
         image.read(fnImg,true,-1,true);
         image().setXmippOrigin();
